Fixed the Twitch Command
