+++ src/FlawBOT.Core/FlawBOT.Core.csproj
@@ -11,7 +11,6 @@
     <LangVersion>7.2</LangVersion>
     <AssemblyVersion>2.4.0.0</AssemblyVersion>
     <FileVersion>2.4.0.0</FileVersion>
-    <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
     <AssemblyName>FlawBOT.Core</AssemblyName>
     <RootNamespace>FlawBOT.Core</RootNamespace>
     <PackageTags />
