--- src/FlawBOT.Core/Properties/launchSettings.json
@@ -1,10 +1,0 @@
-{
-  "profiles": {
-    "FlawBOT.Core": {
-      "commandName": "Project"
-    },
-    "Docker": {
-      "commandName": "Docker"
-    }
-  }
-}
