+++ src/FlawBOT.Framework/Models/Search/TwitchData.cs
@@ -6,166 +6,64 @@ namespace FlawBOT.Framework.Models
 {
     public class TwitchData
     {
-        [JsonProperty("users")]
-        public List<User> Users { get; set; }
-
-        //[JsonProperty("stream")]
-        //public Stream Stream { get; set; }
-    }
-
-    public class User
-    {
-        [JsonProperty("_id")]
-        public string ID { get; set; }
-
-        [JsonProperty("bio")]
-        public string Bio { get; set; }
-
-        [JsonProperty("created_at")]
-        public DateTime CreationDate { get; set; }
-
-        [JsonProperty("display_name")]
-        public string DisplayName { get; set; }
-
-        [JsonProperty("logo")]
-        public string Logo { get; set; }
-
-        [JsonProperty("name")]
-        public string Name { get; set; }
-
-        [JsonProperty("type")]
-        public string UserType { get; set; }
-
-        [JsonProperty("updated_at")]
-        public DateTime UpdateDate { get; set; }
+        [JsonProperty("data")]
+        public List<Stream> Stream { get; set; }
     }
 
     public class Stream
     {
-        [JsonProperty("game")]
-        public string Game { get; set; }
+        [JsonProperty("id")]
+        public int ID { get; set; }
 
-        [JsonProperty("viewers")]
-        public int Viewers { get; set; }
+        [JsonProperty("user_id")]
+        public int UserID { get; set; }
 
-        [JsonProperty("average_fps")]
-        public double AverageFPS { get; set; }
+        [JsonProperty("user_name")]
+        public string UserName { get; set; }
 
-        [JsonProperty("delay")]
-        public int Delay { get; set; }
+        [JsonProperty("game_id")]
+        public int GameID { get; set; }
 
-        [JsonProperty("created_at")]
-        public DateTime CreatedAt { get; set; }
+        [JsonProperty("type")]
+        public string Type { get; set; }
 
-        [JsonProperty("stream_type")]
-        public string StreamType { get; set; }
+        [JsonProperty("title")]
+        public string Title { get; set; }
 
-        [JsonProperty("channel")]
-        public Channel Channel { get; set; }
-    }
+        [JsonProperty("viewer_count")]
+        public int ViewCount { get; set; }
 
-    public class Channel
-    {
-        [JsonProperty("mature")]
-        public bool IsMature { get; set; }
-
-        [JsonProperty("partner")]
-        public bool IsPartner { get; set; }
-
-        [JsonProperty("status")]
-        public string Status { get; set; }
-
-        [JsonProperty("broadcaster_language")]
-        public string BroadcasterLanguage { get; set; }
-
-        [JsonProperty("broadcaster_software")]
-        public string Software { get; set; }
-
-        [JsonProperty("display_name")]
-        public string DisplayName { get; set; }
-
-        [JsonProperty("game")]
-        public string Game { get; set; }
+        [JsonProperty("started_at")]
+        public string StartTime { get; set; }
 
         [JsonProperty("language")]
         public string Language { get; set; }
 
-        [JsonProperty("_id")]
-        public int Id { get; set; }
-
-        [JsonProperty("name")]
-        public string Name { get; set; }
-
-        [JsonProperty("created_at")]
-        public DateTime CreatedAt { get; set; }
-
-        [JsonProperty("updated_at")]
-        public DateTime UpdatedAt { get; set; }
-
-        [JsonProperty("delay")]
-        public object Delay { get; set; }
-
-        [JsonProperty("logo")]
-        public string Logo { get; set; }
-
-        [JsonProperty("banner")]
-        public object Banner { get; set; }
-
-        [JsonProperty("video_banner")]
-        public string VideoBanner { get; set; }
-
-        [JsonProperty("background")]
-        public object Background { get; set; }
-
-        [JsonProperty("profile_banner")]
-        public string ProfileBanner { get; set; }
-
-        [JsonProperty("profile_banner_background_color")]
-        public string ProfileBannerColor { get; set; }
-
-        [JsonProperty("url")]
-        public string Url { get; set; }
-
-        [JsonProperty("views")]
-        public int Views { get; set; }
-
-        [JsonProperty("followers")]
-        public int Followers { get; set; }
-
-        [JsonProperty("_links")]
-        public Links Links { get; set; }
+        [JsonProperty("thumbnail_url")]
+        public string ThumbnailUrl { get; set; }
     }
 
-    public class Links
+    public class Streamer
     {
-        [JsonProperty("self")]
-        public string Self { get; set; }
+        [JsonProperty("id")]
+        public int ID { get; set; }
 
-        [JsonProperty("follows")]
-        public string Follows { get; set; }
+        [JsonProperty("login")]
+        public string Login { get; set; }
 
-        [JsonProperty("commercial")]
-        public string Commercial { get; set; }
+        [JsonProperty("display_name")]
+        public string DisplayName { get; set; }
 
-        [JsonProperty("stream_key")]
-        public string StreamKey { get; set; }
+        [JsonProperty("description")]
+        public string Description { get; set; }
 
-        [JsonProperty("chat")]
-        public string Chat { get; set; }
+        [JsonProperty("profile_image_url")]
+        public string ProfileImageUrl { get; set; }
 
-        [JsonProperty("features")]
-        public string Features { get; set; }
+        [JsonProperty("offline_image_url")]
+        public string OfflineImageUrl { get; set; }
 
-        [JsonProperty("subscriptions")]
-        public string Subscriptions { get; set; }
-
-        [JsonProperty("editors")]
-        public string Editors { get; set; }
-
-        [JsonProperty("teams")]
-        public string Teams { get; set; }
-
-        [JsonProperty("videos")]
-        public string Videos { get; set; }
+        [JsonProperty("view_count")]
+        public int ViewCount { get; set; }
     }
 }
