+++ src/FlawBOT.Framework/Models/Server/BotData.cs
